+++ include/vrv/object.h
@@ -220,6 +220,7 @@ public:
     ///@{
     int GetChildCount() const { return (int)m_children.size(); }
     int GetChildCount(const ClassId classId) const;
+    int GetChildCount(const ClassId, int deepth);
     ///@}
 
     /**