+++ include/vrv/staffdef.h
@@ -54,8 +54,8 @@ public:
      * @name Setter and getter of the drawing visible flag
      */
     ///@{
-    VisibilityOptimization GetDrawingVisibility() const { return m_drawingIsVisible; }
-    void SetDrawingVisibility(VisibilityOptimization drawingIsVisible) { m_drawingIsVisible = drawingIsVisible; }
+    VisibilityOptimization GetDrawingVisibility() const { return m_drawingVisibility; }
+    void SetDrawingVisibility(VisibilityOptimization drawingIsVisible) { m_drawingVisibility = drawingIsVisible; }
     ///@}
 
     //----------//
@@ -83,7 +83,7 @@ private:
      * when the corresponding staves contain only mRest elements.
      * By default the value is OPTIMIZATION_NONE
      */
-    VisibilityOptimization m_drawingIsVisible;
+    VisibilityOptimization m_drawingVisibility;
 };
 
 } // namespace vrv