+++ src/view_page.cpp
@@ -147,7 +147,7 @@ void View::DrawSystem(DeviceContext *dc, System *system)
     dc->StartGraphic(system, "", system->GetUuid());
 
     // Draw system divider (from the second one) if scoreDef is optimized
-    if ((system->GetIdx() > 0) && (system->GetDrawingScoreDef()->GetOptimize() != BOOLEAN_false)) {
+    if ((system->GetIdx() > 0) && system->IsDrawingOptimized()) {
         int x1 = system->GetDrawingX() - m_doc->GetDrawingUnit(100) * 3;
         int y1 = system->GetDrawingY() - m_doc->GetDrawingUnit(100) * 1;
         int x2 = system->GetDrawingX() + m_doc->GetDrawingUnit(100) * 3;